Fix dir creation for single and multi host disk use cases
(SOFTWARE-4159)

For the multi-disk case, we can be prescriptive and tell users where
to mount their data, meta, and namespace (i.e. oss.localroot ) disks,
so we can create the necessary directories in the image
itself. N.B. oss.space directives will create the meta/data dirs
in the paths specified in the second argument so we only have to worry
about creating the namespace dir. This works nicely for the ATLAS
XCache case since they use multiple disks and they're overriding our
supervisor usage so we need this config baked into the image itself.

In the single-disk case, users should be allowed to mount their disk
to /xcache and get everything they need (meta, data, namespace)
without having to worry about having an empty disk to begin
with. In this case, the host dir's contents would override anything
that we create in /xcache so we need to create the namespace dir after
container startup, so we do this in a supervisor startup script.